+++ lib/autoupdate/agent/config.go
@@ -21,13 +21,12 @@ package agent
 import (
 	"errors"
 	"fmt"
+	"io"
 	"io/fs"
 	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
-	"github.com/google/renameio/v2"
 	"github.com/gravitational/trace"
 	"gopkg.in/yaml.v3"
 
@@ -198,21 +197,9 @@ func readConfig(path string) (*UpdateConfig, error) {
 
 // writeConfig writes UpdateConfig to a file atomically, ensuring the file cannot be corrupted.
 func writeConfig(filename string, cfg *UpdateConfig) error {
-	opts := []renameio.Option{
-		renameio.WithPermissions(configFileMode),
-		renameio.WithExistingPermissions(),
-		renameio.WithTempDir(filepath.Dir(filename)),
-	}
-	t, err := renameio.NewPendingFile(filename, opts...)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	defer t.Cleanup()
-	err = yaml.NewEncoder(t).Encode(cfg)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	return trace.Wrap(t.CloseAtomicallyReplace())
+	return trace.Wrap(writeAtomicWithinDir(filename, configFileMode, func(w io.Writer) error {
+		return trace.Wrap(yaml.NewEncoder(w).Encode(cfg))
+	}))
 }
 
 func validateConfigSpec(spec *UpdateSpec, override OverrideConfig) error {
