+++ lib/autoupdate/agent/updater.go
@@ -35,7 +35,6 @@ import (
 	"path/filepath"
 	"runtime"
 	"slices"
-	"syscall"
 	"time"
 
 	"github.com/google/uuid"
@@ -84,13 +83,6 @@ const (
 var (
 	initTime = time.Now().UTC()
 )
-
-// SetRequiredUmask sets the umask to match the systemd umask that the teleport-update service will execute with.
-// This ensures consistent file permissions.
-// NOTE: This must be run in main.go before any goroutines that create files are started.
-func SetRequiredUmask(ctx context.Context, log *slog.Logger) {
-	warnUmask(ctx, log, syscall.Umask(requiredUmask))
-}
 
 func warnUmask(ctx context.Context, log *slog.Logger, old int) {
 	if old&^requiredUmask != 0 {
